+++ src/app/templates/scripts/analyze.sh
@@ -2,6 +2,6 @@
 
 set -e
 
-luau-lsp analyze --base-luaurc=.luaurc --settings=.luau-analyze.json \
+luau-lsp analyze --base-luaurc=.luaurc \
     --ignore '**/node_modules/**' --ignore 'node_modules/**' \
     <%- root %>