+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.1.6
+
 - update darklua configurations to use `.luaurc` instead of providing custom aliases ([#44](https://github.com/seaofvoices/generator-luau/pull/44))
 - update luau-lsp VSCode settings and remove `.luau-analyze.json` file since it is not needed anymore ([#43](https://github.com/seaofvoices/generator-luau/pull/43))
 - change default file extension to `.luau` ([#41](https://github.com/seaofvoices/generator-luau/pull/41))